requirements.txt
Django
djangorestframework
django-filter
Pillow
django-cors-headers
channels
channels_redis
djangorestframework-simplejwt
django-two-factor-auth
django-otp
psycopg2
redis
daphne
drf-yasg
python-socketio
python-dotenv
requests